Experiments cleanup
